requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.4
django-dotenv==1.3.0
django-model-utils==2.2
django-toolbelt==0.0.1
djangorestframework==3.0.5
gunicorn==19.2.1
psycopg2==2.6
static3==0.5.1
